requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Deprecated==1.2.18
distro==1.9.0
Django==5.2.4
djangorestframework==3.16.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
Markdown==3.8.2
netaddr==1.3.0
openai==1.97.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyOpenSSL==25.1.0
python-dotenv==1.1.1
requests==2.32.4
service-identity==24.2.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
wrapt==1.17.2
yookassa==3.6.0
zope.interface==7.2
